requirements.txt
# Discord Bot Dependencies
discord.py==2.3.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic>=2.10.0
lxml>=5.0.0